Refactor event category components and move to new module structure
- Deleted old event-related components: EventsFeed, EventsList, EventsStatsGrid, EventsTableColumns, EventsTabsSection.
- Created new components under the `src/modules/category/ui` directory for better organization.
- Introduced `CategoryDetailsView` to encapsulate the event stats, chart, and feed.
- Implemented `CategoryTimeRangeTabs` for time range selection.
- Updated hooks and types to reflect the new module structure.
- Enhanced event data handling with new data table and list components.
- Improved loading states and empty states for better user experience.
