+++ src/modules/category/ui/sections/events-stats-grid.tsx
@@ -2,7 +2,7 @@
 
 import { Card, CardContent } from "@/components/ui/card";
 import { BarChart } from "lucide-react";
-import { useEventCategoryParams } from "@/hooks/use-event-category-params";
+import { useEventCategoryParams } from "@/modules/category/hooks/use-event-category-params";
 import { useQuery } from "@tanstack/react-query";
 import { useTRPC } from "@/trpc/client";
 import { EventCategory } from "@/db/schema";
